requirements.txt
numpy
faiss-cpu
sentence-transformers
openai
python-dotenv
tqdm
sqlite-utils
requests

tiktoken; extra == "tiktoken"